refactor: remove scaling benchmark, superseded by inference_speed
Delete the scaling benchmark and all references to it. inference_speed
is now self-contained (inlines the constants/helpers it used to import
from scaling) and owns the dataset directly (drops data_name, so it
fetches inference_speed.zip). Test fixtures renamed accordingly.

NOTE: the HF archive scaling.zip in InstaDeepAI/MLIPAudit-data must be
renamed to inference_speed.zip or real runs fail to download data.
