+++ CHANGELOG.md
@@ -2,15 +2,16 @@
 
 ## Unreleased
 
-- Add an `inference_speed` benchmark that measures, per system size, both **model
-  throughput** (the raw forward pass — pure network for mlip models, forced ASE
-  recompute for external models) and **MD throughput** on each supported backend
-  (JAX-MD and ASE for mlip models, ASE for external), so JAX and external models can be
-  compared on the shared ASE backend. It reuses the `scaling` dataset (via the
-  base-class `data_name`/`data_dir`) and produces a (hardware-relative) speed score
-  from the per-atom model forward time, contributing to the overall model score. The
-  GUI switches between metrics/backends on log–log axes with power-law fits and
-  variance, and a summary table including the model's graph cutoff.
+- Replace the `scaling` benchmark with a new `inference_speed` benchmark that measures,
+  per system size, both **model throughput** (the raw forward pass — pure network for
+  mlip models, forced ASE recompute for external models) and **MD throughput** on each
+  supported backend (JAX-MD and ASE for mlip models, ASE for external), so JAX and
+  external models can be compared on the shared ASE backend. Unlike `scaling`, it
+  produces a (hardware-relative) speed score from the per-atom model forward time,
+  contributing to the overall model score. The GUI switches between metrics/backends on
+  log–log axes with power-law fits and variance, and a summary table including the
+  model's graph cutoff. The size-stratified protein dataset is renamed from `scaling` to
+  `inference_speed` (the corresponding HuggingFace archive must be renamed accordingly).
 
 ## Release 0.1.5
 
@@ -27,13 +28,6 @@
   model conserves the total energy, quantified by the drift of the total energy over
   the trajectory.
 - Require `mlip>=0.2.3` and adapt `ASESimulationEngineWithCalculator` to its refactor
-- Add an `inference_speed` benchmark that measures, per system size, both **model
-  throughput** (the raw forward pass — pure network for mlip models, forced ASE
-  recompute for external models) and **MD throughput** (end-to-end ns/day). It reuses
-  the `scaling` dataset and produces a (hardware-relative) speed score from the per-atom
-  model forward time, contributing to the overall model score. The GUI switches between
-  metrics on log–log axes with power-law fits and variance, and a summary table
-  including the model's graph cutoff.
 - Benchmarks can now reuse another benchmark's dataset via the `dataset_name` attribute.
   of `ASESimulationEngine._init_box` into the `resolve_atoms_cell` helper.
 - Replace the orexin-beta system with the villin headpiece (PDB `1UNC`) in the
