--- docs/source/api_reference/general/scaling.rst
@@ -1,20 +1,0 @@
-.. _scaling_api:
-
-Scaling
-=========
-
-.. module:: mlipaudit.benchmarks.scaling.scaling
-
-.. autoclass:: ScalingBenchmark
-
-    .. automethod:: __init__
-
-    .. automethod:: run_model
-
-    .. automethod:: analyze
-
-.. autoclass:: ScalingResult
-
-.. autoclass:: ScalingStructureResult
-
-.. autoclass:: ScalingModelOutput
