+++ docs/source/api_reference/index.rst
@@ -42,6 +42,5 @@ Benchmark implementations
     biomolecules/folding_stability
     biomolecules/sampling
     general/stability
-    general/scaling
     general/nve_energy_conservation
     general/inference_speed
