+++ docs/source/benchmarks/general/index.rst
@@ -10,6 +10,5 @@ applicable across molecular systems.
     :maxdepth: 1
 
     Stability <stability>
-    Scaling <scaling>
     NVE energy conservation <nve_energy_conservation>
     Inference Speed <inference_speed>
